% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSI_FingerID_writer.R
\name{CSI_FingerID_writer}
\alias{CSI_FingerID_writer}
\title{Writing the spectral library into files compatible with CSI:FingerID (SIRIUS 4.0)}
\usage{
CSI_FingerID_writer(library)
}
\arguments{
\item{library}{A list generated by the function library_generator() or the name of mgf spectral library file}
}
\description{
The function creates a folder "CSI_FingerID" where .ms files are stored. These files can be imported to SIRIUS 4.0 software for structure elucidation. The files are named by IDs in user-defined metadata and contain at least MS2 scan. MS1 scan is added if available.
}
