% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library_generator.r
\name{library_generator}
\alias{library_generator}
\title{Batch processing of LC-MS/MS chromatograms into a spectral library}
\usage{
library_generator(raw_data_files, metadata_file, mslevel = c(1, 2),
  MS2_type = "DDA", rt_search = 12, ppm_search = 20, search_adduct = T,
  baseline = 1000, normalized = T, input_library = "",
  output_library = "")
}
\arguments{
\item{raw_data_files}{A character vector of file names of chromatograms from which scans are extracted. All files must have be in centroid-mode with mzML or mzMXL extension!}

\item{metadata_file}{File name of the metadata. Must be a single character with csv extension.}

\item{mslevel}{Must be 1 (if only MS1 scans/isotopic patterns of targeted m/z are extracted), 2 (if only MS2 scans are extracted) or c(1,2) (if both MS1 and MS2 scans are extracted)}

\item{MS2_type}{A single character ("DDA" or "Targeted") if all raw_dat_files are acquired in the same mode; A character vector precising the acquisition mode of each file in raw_data_files (e.g. c("DDA","Targeted","DDA"))}

\item{rt_search}{Retention time search tolerance (in second) for targeted RT}

\item{ppm_search}{m/z search tolerance (in ppm) for targeted m/z}

\item{search_adduct}{Logical indicating whether [M+Na+] (in positive mode) or [M+Cl-] (in negative mode) adducts are searched. If FALSE, only [M+H+] and [M-H+] adducts are considered.}

\item{baseline}{Numeric, the minimum intensity that is considered as a mass peak and written into the library}

\item{normalized}{Logical, TRUE if the intensities of extracted spectra need to normalized so that the intensity of highest peak will be 100}

\item{input_library}{Name of the library into which new scans are added, the file extension must be mgf; please set to empty string "" if the new library has no dependency with previous ones.}

\item{output_library}{Name of the output library, the file extension must be mgf}
}
\value{
\itemize{
  \item{"sp" ~ List of all extracted spectra. Each spectrum is a data matrix with two columns: m/z and intensity}
  \item{"metadata" ~ Data frame containing metadata of extracted scans. PEPMASS and RT are updated based on actually-detected scans. Following columns are added: FILENAME, MSLEVEL, TIC, ADDUCT, SCANNUMBER and SCANS}
  \item{"<ouput_library>" ~ A mgf spectral library file will be written in user's working directory. It contains both spectra and metadata}
  \item{"<ouput_library.txt>" ~ Metadata will be written as a tab-seperated .txt file in user's working directory. Users can check this file in excel or open office.}
}
}
\description{
The function picks up scans according to m/z (and retention time) specified in the metadata and merge them into a spcetral library (new or existing). The raw LC-MS/MS files must be centroid-mode mzML, mzMXL or mzData
}
\examples{
### We download our three test data sets:
# Details of these data can be found at: https://zenodo.org/record/1322562

url = "https://zenodo.org/record/1322562/files/"
original_files = c("TESTMIX2_180504_MAS011_06.mzXML","JNJ42165279_171214_MAS006_14.mzXML","GMP_R601592_150925_MAS006_04.mzXML")
download.file(paste0(url,original_files[1]),destfile="F1.mzXML")
download.file(paste0(url,original_files[2]),destfile="F2.mzXML")
download.file(paste0(url,original_files[3]),destfile="F3.mzXML")
### It is time to batch-process the first two files and create our first spectral library:

raw_data_files=c("F1.mzXML","F2.mzXML")
metadata_file = paste0(url,"library_metadata.csv")
mslevel = c(1,2) # Both MS1 and MS2 scans are extracted! MS1 scan contains isotopic pattern of targeted m/z
MS2_type = c("DDA","Targeted") # Mode of MS/MS experiment.
rt_search = 12 # Retention time tolerance (s)
ppm_search = 10  # Mass tolerance (ppm)
baseline = 1000 # Noise level of mass spectra, can be determined using spectral visualizer such as MZMine
input_library = "" # A brand new library, there's no previous dependency
output_library = "library_V1.mgf" # Name of the library

library1=library_generator(raw_data_files,metadata_file,mslevel,MS2_type,rt_search,ppm_search,search_adduct = F,baseline,normalized = T, input_library, output_library)

### We added the targeted scans of F3.mzXML to spectral library version 2:

raw_data_files="F3.mzXML" # The new LC-MS/MS data
MS2_type = "DDA"
metadata_file = paste0(url,"library_metadata_GMP.csv")
input_library = "library_V1.mgf" # The first mgf file of library1
output_library = "library_V2.mgf" # The name of the new spectral library

library2=library_generator(raw_data_files,metadata_file,mslevel,MS2_type,rt_search,ppm_search,search_adduct = F,baseline,normalized = T, input_library, output_library)

# In the end, two spectral library files "library_V1.mgf" and "library_V2.mgf" should appear in the working directory along with metadata table (txt files)

}
\author{
Youzhong Liu, \email{Youzhong.Liu@uantwerpen.be}
}
