% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library_visualizer.R
\name{library_visualizer}
\alias{library_visualizer}
\title{Visualize selected mass spectra in the spectral library}
\usage{
library_visualizer(library, max.plot = 0, png.out = F)
}
\arguments{
\item{library}{A list generated by the function library_generator() or the name of mgf spectral library file}

\item{max.plot}{An integer. At least 1, maximal number of plots per page}

\item{png.out}{Boolean. True if plotted spectra are exported as png images!}
}
\description{
The function plots all mass spectra in a library
}
\examples{
# Search library using query command lines:
query = library_manager(library1,query=c("IONMODE=Positive","MSLEVEL=2","RT=1.2"), logical="AND", rt_search=6)

# Visualize scans found:
library_visualizer(query$SELECTED)

}
