% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_library.R
\name{match_fragments}
\alias{match_fragments}
\title{Search fragments in the spectral library}
\usage{
match_fragments(library, fragments = NULL, match = c("All", "Partial"),
  ppm_search = 20)
}
\arguments{
\item{library}{A list generated by the function library_generator() or the name of mgf spectral library file}

\item{fragments}{Numeric vector. The m/z value(s) of query fragments.}

\item{match}{Character. "All" if the extracted scan must contain all query fragments; "Partial" if at least one fragment is present.}

\item{ppm_search}{m/z search tolerance (in ppm) for fragments}
}
\value{
\itemize{
 \item{"ID" ~ Vector of characters. ID(s) of metabolic feature(s) that contain query fragments.}
 \item{"SCANS" ~ Numeric vector. Scan number(s) that contain query fragments.}
}
}
\description{
The function searches and plots all mass spectra that contain certain fragments
}
\examples{
matched = match_fragments(library1,c(392.23,450.06))
matched = match_fragments(library1,101.06,ppm_search = 10)

}
