% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_editor.R
\name{meta_editor}
\alias{meta_editor}
\title{Create additional metadata for library generation}
\usage{
meta_editor(ref, adducts = c("Default", "M+H", "M+Na", "M+K", "M+NH4",
  "M-H", "M+Cl"), max.charge = 1)
}
\description{
Function used by library_generator() to create adducts and multiple charged features
}
