% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetFrag_writer.R
\name{MetFrag_writer}
\alias{MetFrag_writer}
\title{Writing the spectral library into files that allow batch processing on Metfrag}
\usage{
MetFrag_writer(library, ppm_search = 20)
}
\arguments{
\item{library}{A list generated by the function library_generator() or the name of mgf spectral library file}

\item{ppm_search}{m/z search tolerance (in ppm) for precursor and fragments}
}
\value{
Please click on "Metfrag_batch_windows.bat" for Metfrag batch-identification!
}
\description{
The function creates a folder "MetFrag" where .txt and .bat files are stored.
}
\examples{
MetFrag_writer("library_V2.mgf",ppm_search = 10)
}
