% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSFinder_writer.R
\name{MSFinder_writer}
\alias{MSFinder_writer}
\title{Writing the spectral library into files compatible with MSFinder}
\usage{
MSFinder_writer(library)
}
\arguments{
\item{library}{A list generated by the function library_generator() or the name of mgf spectral library file}
}
\description{
The function creates a folder "MSFinder" where .mat files are stored. These files can be imported to MSFinder software for structure elucidation. The files are named by IDs in user-defined metadata and contain at least one MS2 scan. The MS1 scan is added if available.
}
\examples{
MSFinder_writer("library_V1.mgf")

}
