% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_library.R
\name{reduce_library_by_id}
\alias{reduce_library_by_id}
\title{Extracting representative scans for each feature (ID) and create a new library}
\usage{
reduce_library_by_id(library, method = c("consensus", "common",
  "max_tic", "max_nb_peaks"), consensus_window = 10,
  output_library = "")
}
\arguments{
\item{library}{A list generated by the function library_generator() or the name of mgf spectral library file}

\item{output_library}{Name of the output library, the file extension must be mgf}

\item{consensus}{Logical. TRUE if consensus spectrum is generated by spectral alignment. FALSE if users only wish to keep the spectrum with the highest TIC.ATTENTION: consensus spectra can be generated only if mass spectra are normalized !}

\item{ppm_window}{m/z tolerance window (in ppm) for spectra alignment}

\item{strict}{Logical. TRUE if only keep mass peaks present in all spectra of an ID}
}
\value{
The same as library_generator()
}
\description{
The function picks a single MS1 and MS2 scan for each ID in the spectral library according to the highest TIC. To generate consensus spectra, all mass spectra of the same ID at the same MS level are aligned (on m/z) and intensities are averaged
}
\examples{

# Withholding the scan (MS1 and MS) with highest TIC for each ID:

library2_1=reduce_library_by_id(library2, consensus=F, output_library="library_V2_filtered.mgf")

# Generating consensus spectra for all scans (MS1 and MS) of the same ID. The scan with highest TIC is used for metadata:

library2_2=reduce_library_by_id(library2, consensus=T, ppm_window = 20, output_library="library_V2_consensus.mgf")

}
\author{
Youzhong Liu, \email{Youzhong.Liu@uantwerpen.be}
}
