% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.spectra.R
\name{visualize.spectra}
\alias{visualize.spectra}
\title{Visualize mass spectra in the spectral library}
\usage{
visualize.spectra(library, ID)
}
\arguments{
\item{library}{A list generated by the function library_generator() or the name of mgf spectral library file}

\item{ID}{Numeric or character. The ID of metabolic feature for which mass spectra will be plotted}
}
\description{
The function plots all mass spectra of a metabolic feature
}
\examples{
visualize.spectra(library1,2)

}
