% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library_generator.r
\name{library_generator}
\alias{library_generator}
\title{Batch processing of mzML/mzXML files into a spectral library}
\usage{
library_generator(raw_data_files, metadata_file, mslevel = c(1, 2),
  MS2_type = "DDA", rt_search = 12, ppm_search = 20, search_adduct = T,
  baseline = 1000, normalized = T, input_library = "",
  output_library = "")
}
\arguments{
\item{raw_data_files}{A character vector of file names of chromatograms from which scans are extracted. All files must have be in centroid-mode with mzML or mzMXL extension!}

\item{metadata_file}{File name of the metadata. Must be a single character with csv extension.}

\item{mslevel}{Must be 1 (if only MS1 scans are extracted), 2 (if only MS2 scans are extracted) or c(1,2) (if both MS1 and MS2 scans are extracted)}

\item{MS2_type}{A single character ("DDA" or "Targeted") if all raw_dat_files are acquired in the same mode; A character vector precising the acquisition mode of each file in raw_data_files (e.g. c("DDA","Targeted","DDA"))}

\item{rt_search}{Retention time search tolerance (in second) for targeted RT}

\item{ppm_search}{m/z search tolerance (in ppm) for targeted m/z}

\item{search_adduct}{Boolean indicating whether [M+Na+] (in positive mode) or [M+Cl-] (in negative mode) adducts are searched}

\item{baseline}{Numeric, the minimum intensity that is considered as a mass peak and written into the library}

\item{input_library}{Name of the library into which new scans are added, the file extension must be mgf; please set to empty string "" if the new library has no dependency with previous ones.}

\item{ouput_library}{Name of the output library, the file extension must be mgf}
}
\value{
\itemize{
  \item{"sp"}{List of all extracted spectra. Each spectrum is a data matrix with two columns: m/z and intensity}
  \item{"metadata"}{Data frame containing metadata of extracted scans. PEPMASS and RT are updated based on actually-detected scans. Following columns are added: FILENAME, MSLEVEL, TIC, ADDUCT, SCANNUMBER and SCANS}
  \item{"<ouput_library>"}{A mgf spectral library file will be written in user's working directory. It contains both spectra and metadata}
  \item{"<ouput_library>.txt"}{Metadata will be written as a .txt file in user's working directory. Users can check this file in excel or open office.}
}
}
\description{
The function picks up scans according to m/z (and retention time) specified in the metadata and merge them into a spcetral library (new or existing). The raw LC-MS/MS files must be centroid-mode mzML or mzMXL
}
\examples{



}
\author{
Youzhong Liu, \email{Youzhong.Liu@uantwerpen.be}
}
