% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library_similarity.R
\name{library_similarity}
\alias{library_similarity}
\title{Matching query spectrum to existing library}
\usage{
library_similarity(library, query_spectrum = NULL, method = c("Simple",
  "Cosinus"), tops = 3, prec_mz = 1000, ppm_search = 20,
  relative = 1)
}
\description{
Matching query spectrum to existing library
}
