% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library_visualizer.R
\name{library_visualizer}
\alias{library_visualizer}
\title{Visualize selected mass spectra in the spectral library}
\usage{
library_visualizer(query, query.mode = c("found", "remained"),
  max.plot = 1)
}
\arguments{
\item{query}{A list of four objects SELECTED_SCANS, LEFT_SCANS, SELECTED_LIBRARY and LEFT_LIBRARY. It should be the output of library_manager().}

\item{query.mode}{A character. Either "found" or "remained".}

\item{max.plot}{An integer. At least 1, maximal number of plots per page}
}
\description{
The function plots all mass spectra found by library query
}
\examples{
# Search library using query command lines:
query = library_manager(library1,query=c("IONMODE=Positive","MSLEVEL=2","RT=1.2"), logical="AND", rt_search=6)

# Visualize scans found:
library_visualizer(query, query.mode="found")

}
