% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_library.R
\name{process_library}
\alias{process_library}
\title{For each feature (ID) extract representative scans}
\usage{
process_library(spectrum_list, metadata, consensus = T, ppm_window = 10,
  clean = T)
}
\arguments{
\item{spectrum_list}{Spectra list of spectral library}

\item{metadata}{Metadata of spectral library}

\item{consensus}{Logical. TRUE if consensus spectrum is generated by spectral alignment. FALSE if users only wish to keep the spectrum with the highest TIC.}

\item{ppm_window}{m/z tolerance window (in ppm) for spectra alignment}

\item{clean}{Logical. TRUE if peaks not present in all spectra behind an ID are removed}
}
\value{
\itemize{
  \item{"sp" ~ List of all extracted spectra. Each spectrum is a data matrix with two columns: m/z and intensity}
  \item{"metadata" ~ Data frame containing metadata of extracted scans. PEPMASS and RT are updated based on actually-detected scans.}
 }
}
\description{
The function picks a single MS1 and MS2 scan for each ID in the spectral library according to the highest TIC. It can also combine all mass spectra behind each ID into one consensus spectrum (by alignment of m/z and averaging of intensity)
}
