% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.spectra.R
\name{visualize.spectra}
\alias{visualize.spectra}
\title{Visualize mass spectra in the spectral library}
\usage{
visualize.spectra(library, ID = NULL, mslevel = c(1, 2), scan = NULL)
}
\arguments{
\item{library}{A list generated by the function library_generator() or the name of mgf spectral library file}

\item{ID}{Numeric or character. The ID of metabolic feature for which mass spectra will be plotted}

\item{mslevel}{Must be 1 (if only MS1 scans are plotted), 2 (if only MS2 scans are plotted) or c(1,2) (if both are plotted)}

\item{scan}{Numeric. Scan number of the spectra plotted. ID or mslevel is not necessary if scan number is precised.}
}
\description{
The function plots all mass spectra of a metabolic feature
}
\examples{
visualize.spectra(library1, ID = 2, mslevel = 2)
visualize.spectra(library1, scan = 1)

}
